Split wrap and table tests
